requirements.txt
asgiref==3.4.1
Django==3.2.9
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-osm-field==0.5.0
django-widget-tweaks==1.4.9
Pillow==8.4.0
pipp==0.0.1
pytz==2021.3
sqlparse==0.4.2
